converted GlobalHeader from js to tsx and hooks
